bulk_prepare_expression {context}	R Documentation

Prepare many expressions

Description

Prepare many expressions

Usage

bulk_prepare_expression(X, FUN, DOTS, do_call, use_names, envir, db)

Arguments

X
Something to iterate over; a vector, list or data.frame (in the case of a data.frame, iteration will be row-by-row)

FUN
A function to apply to each element (or row) of X

DOTS
Additional arguments to apply with each elements of X

do_call
Treat each element of X as a do.call call

use_names
When preparing a data.frame, retain column names as argument names when using do_call. If FALSE then positional matching will be used.

envir
An environment to find variables local to the expression

db
A database to store locals

[Package context version 0.3.0 Index]