bulk_task_save {context}	R Documentation

Save bulk tasks

Description

Save bulk tasks

Usage

bulk_task_save(
  X,
  FUN,
  context,
  DOTS = NULL,
  do_call = FALSE,
  use_names = TRUE,
  envir = parent.frame()
)

Arguments

X
Something to iterate over; a vector, list or data.frame (in the case of a data.frame, iteration will be row-by-row)

FUN
A function to apply to each element (or row) of X

context
A context

DOTS
Additional arguments to apply with each elements of X

do_call
Treat each element of X as a do.call call

use_names
When preparing a data.frame, retain column names as argument names when using do_call. If FALSE then positional matching will be used.

envir
An environment to find variables local to the expression

[Package context version 0.3.0 Index]