context_list {context}	R Documentation

List save contexts

Description

List saved contexts

Usage

context_list(db, named = FALSE, error = TRUE)

context_info(db, error = TRUE)

Arguments

db
Something for which a context database can be created; this can the the path to the context, a context_root object, or a context object.

named
Logical, indicating if the context name should be used to name the output vector.

error
Throw an error if the context database cannot be connected constructed (e.g., if the path given does not exist).

Author(s)

Rich FitzJohn

[Package context version 0.3.0 Index]