context_load {context}	R Documentation

Load a context

Description

Load a context

Usage

context_load(ctx, envir = .GlobalEnv, refresh = FALSE)

Arguments

ctx
A context object, as read by context_read

envir
The environment to source files into

refresh
Refresh the context, even when it has been loaded already? Note that this may not always behave as expected because items not created by sourcing an R file will still be there from previous runs, and packages loaded will not be reloaded in a new order.

[Package context version 0.3.0 Index]