context_log {context}	R Documentation

Send entry to context log

Description

Send an entry to the context log. This is designed primarily for use with packages that build off of context, so that they can log in a consistent way.

Usage

context_log(topic, value)

Arguments

topic
Up to 9 character text string with the log topic

value
Character string with the log entry

[Package context version 0.3.0 Index]