context_root_get {context}	R Documentation

Find context root

Description

Find the context root. Designed for internal use

Usage

context_root_get(root, db = NULL)

Arguments

root
An object; either a character string (interpreted as a path), a context_root object (such as returned by this function) or a list/environment object with a root element that is a context_root object.

db
Optionally, a copy of the storr database (if already opened). Do not specify this unless you definitely have the correct database in hand.

[Package context version 0.3.0 Index]