parallel_cluster_start {context}	R Documentation

Start a sub-cluster

Description

Start a sub-cluster, using the parallel package. This will be available via either the return value of this function, the parallel_cluster function or by using cl = NULL with any of the parallel package functions. The cluster will be started so that it is ready to use the context.

Usage

parallel_cluster_start(n, ctx)

parallel_cluster_stop()

parallel_cluster()

Arguments

n
The number of nodes. No attempt at guessing this number is made as that is terribly error prone. If you're using this function you should know how many resources you have available.

ctx
The context to initialise on each cluster node.

[Package context version 0.3.0 Index]