task_save {context}	R Documentation

Save and reload tasks

Description

Save and reload tasks. Tasks consist of an expression bound to a context.

Usage

task_save(expr, context, envir = parent.frame())

Arguments

expr
An expression to save

context
A context object

envir
Passed through to context_save when locating local variables.

Value

An identifier that can be used to retrieve or run the task later. This is simply a short string.

[Package context version 0.3.0 Index]