task_log {context}	R Documentation

Return task log

Description

Return the log of a task, if enabled.

Usage

task_log(id, root, parse = TRUE)

Arguments

id
Single task identifier

root
A context root (not just the db as in task_result as we need to know the actual path to the root). A context object is also OK.

parse
Parse the log output into a context_log object, which will pretty print and can be more easily inspected. If FALSE then the raw log will be returned as a character vector, one element per line of text

Details

The returned object is of class task_log, which has a print method that will nicely display. Output is grouped into phases.

[Package context version 0.3.0 Index]