task_result {context}	R Documentation

Fetch task result

Description

Fetch result from completed task.

Usage

task_result(id, db, allow_incomplete = FALSE)

Arguments

id
Single task identifier

db
Something that can be converted to a context db object (a database, root or context).

allow_incomplete
Should we avoid throwing an error if a task is not completed? Used internally, and not generally needed.

[Package context version 0.3.0 Index]