task_run {context}	R Documentation

Run a task

Description

Run a task

Usage

task_run(id, context, filename = NULL)

Arguments

id
A task identifier

context
A context object

filename
Filename to log all output to. This will sink the message stream as well as the output stream, so if specified (i.e., is non-NULL) then this function will apparently print no output to the console, which will make debugging more difficult when run interactively. However, when run non-interactively, especially on remote servers, this will allow collection of diagnostics that facilitate debugging.

[Package context version 0.3.0 Index]