task_run_external {context}	R Documentation

Run a task in separate process

Description

Run a task in a separate process, using [callr::r]. Unlike [context::task_run] this does not return the value, and is called for the side effect of writing to the context.

Usage

task_run_external(root, identifier, task_id, path_log)

Arguments

root
Something interpretable as the context root; either

identifier
Either the id or name of a context (see context_list)

task_id
A task identifier

path_log
Path to log file

[Package context version 0.3.0 Index]