task_times {context}	R Documentation

Fetch task times

Description

Fetch times taken to queue, run, and since running a task.

Usage

task_times(ids, db, unit_elapsed = "secs", sorted = TRUE)

Arguments

ids
Vector of task ids

db
Something that can be converted to a context db object (a database, root or context).

unit_elapsed
Elapsed time unit. The default is "secs". This is passed to the as.numeric method of a difftime object.

sorted
Sort the output in terms of submitted time? If FALSE then the output is sorted based on task ids.

Author(s)

Rich FitzJohn

[Package context version 0.3.0 Index]